Fix MSVC warnings and remove disabled warning flags

+++ onnx/common/ir_pb_converter.cc
@@ -679,7 +679,7 @@ void encodeGraph(GraphProto* p_g, const std::shared_ptr<Graph>& g) {
   }
 
   auto num_initializers = g->initializers().size();
-  for (unsigned int i = 0; i < num_initializers; i++) {
+  for (size_t i = 0; i < num_initializers; i++) {
     auto p = p_g->add_initializer();
     p->set_name(g->initializer_names()[i]);
     encodeTensor(p, g->initializers()[i]);
